requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
docker==7.0.0
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
tenacity==8.2.3
